feat: Add react-scroll library for smooth scrolling
This commit adds the `react-scroll` library as a new dependency in the `package.json` file. This library will be used to implement smooth scrolling functionality in the application.

It also updates the email address in the `Contact.tsx` file to `deniz20070206@gmail.com` for better communication with users.

Additionally, a new file `index.ts` is added to the `app/data2` directory, which exports an array of project objects.

The `page.tsx` file is modified to include a new component `RecentProjects2` and import the `navItems` array from the `data` file.

Finally, the `Banner.tsx` file is updated to include an `id` attribute for the contact section and update the email address to `deniz20070206@gmail.com`.

This commit enhances the functionality and content of the application, improving the user experience and providing more project options.