chore: Refactor GET method in route.ts to support pagination
This commit refactors the GET method in the `route.ts` file to support pagination. It introduces a `limit` query parameter to specify the maximum number of projects to retrieve. The projects are sorted in descending order by `_id`. If a limit is provided, it is applied to the query using the `limit` method.

These changes enhance the functionality of the GET method by allowing the client to control the number of projects returned, providing a more efficient and flexible API.